requirements.txt
aiogram
python-decouple
psycopg2
sqlalchemy
sqlalchemy_utils